[IMP] mail_debrand: Split code into model-specific files
